@@ -139,34 +140,34 @@
     "serve-static": "^1.13.1",
     "slash": "~1.0.0",
     "source-map-explorer": "^1.5.0",
-    "supports-color": "^4.4.0",
+    "supports-color": "^5.0.0",
     "systemjs-builder": "0.16.12",
     "temp": "^0.8.3",
     "tildify": "^1.2.0",
     "traceur": "^0.0.111",
     "ts-node": "^3.3.0",
-    "tslint": "^5.7.0",
-    "typescript": "^2.5.3",
+    "tslint": "^5.8.0",
+    "typescript": "^2.6.1",
     "walk": "^2.3.9",
-    "yargs": "^9.0.1"
+    "yargs": "^10.0.3"
   },
   "dependencies": {
-    "@angular/animations": "^4.4.4",
-    "@angular/common": "^4.4.4",
-    "@angular/compiler": "^4.4.4",
-    "@angular/core": "^4.4.4",
-    "@angular/forms": "^4.4.4",
-    "@angular/http": "^4.4.4",
-    "@angular/platform-browser": "^4.4.4",
-    "@angular/platform-browser-dynamic": "^4.4.4",
-    "@angular/router": "^4.4.4",
-    "@angular/service-worker": "^1.0.0-beta.16",
-    "@angular/material": "2.0.0-beta.12",
-    "@angular/cdk": "2.0.0-beta.12",
+    "@angular/animations": "^5.0.0",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/http": "^5.0.0",
+    "@angular/platform-browser": "^5.0.0",
+    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/router": "^5.0.0",
+    "@angular/service-worker": "^5.0.0",
+    "@angular/material": "5.0.0-rc0",
+    "@angular/cdk": "5.0.0-rc0",
     "@types/hammerjs": "^2.0.31",
     "core-js": "^2.5.1",
     "intl": "^1.2.5",
-    "rxjs": "^5.5.0",
+    "rxjs": "^5.5.2",
     "systemjs": "0.20.19",
     "zone.js": "0.8.18",
     "reflect-metadata": "^0.1.8",
