@@ -1,8 +1,7 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { APP_BASE_HREF } from '@angular/common';
-import { HttpModule } from '@angular/http';
-
+import { HttpClientModule } from '@angular/common/http';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
 
@@ -13,7 +12,7 @@ import { MaterialModule } from './shared/material/material.module';
 
 
 @NgModule({
-imports: [BrowserModule, HttpModule, AppRoutingModule, AboutModule, HomeModule, SharedModule.forRoot(),
+imports: [BrowserModule, HttpClientModule, AppRoutingModule, AboutModule, HomeModule, SharedModule.forRoot(),
   MaterialModule],
   declarations: [AppComponent],
   providers: [{
