@@ -19,8 +19,8 @@ export function main() {
       'should work',
       async(() => {
         TestBed.compileComponents().then(() => {
-          let fixture = TestBed.createComponent(TestComponent);
-          let aboutDOMEl = fixture.debugElement.children[0].nativeElement;
+          const fixture = TestBed.createComponent(TestComponent);
+          const aboutDOMEl = fixture.debugElement.children[0].nativeElement;
 
           expect(aboutDOMEl.querySelectorAll('h2')[0].textContent).toEqual(
             'Features'
