@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Http, Response } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
 // import 'rxjs/add/operator/do';  // for debugging
 
@@ -10,11 +10,11 @@ import { Observable } from 'rxjs/Observable';
 export class NameListService {
 
   /**
-   * Creates a new NameListService with the injected Http.
-   * @param {Http} http - The injected Http.
+   * Creates a new NameListService with the injected HttpClient.
+   * @param {HttpClient} http - The injected HttpClient.
    * @constructor
    */
-  constructor(private http: Http) {}
+  constructor(private http: HttpClient) {}
 
   /**
    * Returns an Observable for the HTTP GET request for the JSON resource.
@@ -22,7 +22,6 @@ export class NameListService {
    */
   get(): Observable<string[]> {
     return this.http.get('assets/data.json')
-                    .map((res: Response) => res.json())
     //              .do(data => console.log('server data:', data))  // debug
                     .catch(this.handleError);
   }
@@ -33,7 +32,7 @@ export class NameListService {
   private handleError (error: any) {
     // In a real world app, we might use a remote logging infrastructure
     // We'd also dig deeper into the error to get a better message
-    let errMsg = (error.message) ? error.message :
+    const errMsg = (error.message) ? error.message :
       error.status ? `${error.status} - ${error.statusText}` : 'Server error';
     console.error(errMsg); // log to console instead
     return Observable.throw(errMsg);
