@@ -515,6 +515,7 @@ export class SeedConfig {
         'node_modules/@angular/platform-browser-dynamic/bundles/platform-browser-dynamic-testing.umd.js',
       '@angular/router/testing':
         'node_modules/@angular/router/bundles/router-testing.umd.js',
+      'rxjs/operators': 'node_modules/rxjs/operators/index.js',
 
       'app/': `${this.APP_BASE}app/`,
       // For test config
@@ -561,6 +562,7 @@ export class SeedConfig {
       '@angular/common/http':
         'node_modules/@angular/common/bundles/common-http.umd.js',
       'tslib': 'node_modules/tslib/tslib.js',
+      'rxjs/operators': 'node_modules/rxjs/operators/index.js',
       'dist/tmp/node_modules/*': 'dist/tmp/node_modules/*',
       'node_modules/*': 'node_modules/*',
       '*': 'node_modules/*'
@@ -685,7 +687,7 @@ export class SeedConfig {
      * Example: `npm start -- --b`
      * @return {any}
      */
-    let defaults = {
+    const defaults = {
       'browser-sync': {
         middleware: [
           require('connect-history-api-fallback')({
@@ -853,7 +855,7 @@ function filterDependency(type: string, d: InjectableDependency): boolean {
  * @return {number} The applications version.
  */
 function appVersion(): number | string {
-  var pkg = require('../../package.json');
+  const pkg = require('../../package.json');
   return pkg.version;
 }
 
@@ -861,9 +863,9 @@ function appVersion(): number | string {
  * Returns the application build type.
  */
 function getBuildType() {
-  let type = (argv['build-type'] || argv['env'] || '').toLowerCase();
-  let base: string[] = argv['_'];
-  let prodKeyword = !!base
+  const type = (argv['build-type'] || argv['env'] || '').toLowerCase();
+  const base: string[] = argv['_'];
+  const prodKeyword = !!base
     .filter(o => o.indexOf(BUILD_TYPES.PRODUCTION) >= 0)
     .pop();
   if ((base && prodKeyword) || type === BUILD_TYPES.PRODUCTION) {
@@ -874,6 +876,6 @@ function getBuildType() {
 }
 
 function getSmeOutFormat() {
-  let format = (argv['sme-out-format'] || '').toUpperCase();
+  const format = (argv['sme-out-format'] || '').toUpperCase();
   return SME_OUTPUT_FORMATS[format] || SME_OUTPUT_FORMATS.HTML;
 }
