@@ -4,7 +4,7 @@ import { join } from 'path';
 import Config from '../../config';
 
 // TODO There should be more elegant to prevent empty directories from copying
-var onlyDirs = function (es: any) {
+const onlyDirs = function (es: any) {
   return es.map(function (file: any, cb: any) {
     if (file.stat.isFile()) {
       return cb(null, file);
@@ -19,7 +19,7 @@ var onlyDirs = function (es: any) {
  * `dist/prod` directory.
  */
 export = () => {
-  let es: any = require('event-stream');
+  const es: any = require('event-stream');
   return gulp.src([
     join(Config.APP_SRC, '**'),
     join(Config.NPM_BASE, '@angular', 'service-worker', 'bundles', 'worker-basic.min.js'),
