@@ -60,7 +60,7 @@ function injectCss() {
  */
 function transformPath() {
   return function(filepath: string) {
-    let path: Array<string> = normalize(filepath).split(sep);
+    const path: Array<string> = normalize(filepath).split(sep);
     let slice_after = path.indexOf(Config.APP_DEST);
     if (slice_after > -1) {
       slice_after++;
