@@ -12,12 +12,12 @@ const plugins = <any>gulpLoadPlugins();
  * for the e2e environment.
  */
 export = () => {
-  let tsProject = makeTsProject({ target: 'es2015' }, Config.E2E_SRC);
-  let src = [
+  const tsProject = makeTsProject({ target: 'es2015' }, Config.E2E_SRC);
+  const src = [
     Config.TOOLS_DIR + '/manual_typings/**/*.d.ts',
     join(Config.E2E_SRC, '**/*.ts')
   ];
-  let result = gulp
+  const result = gulp
     .src(src)
     .pipe(plugins.plumber())
     .pipe(plugins.sourcemaps.init())
