@@ -13,8 +13,8 @@ const plugins = <any>gulpLoadPlugins();
  */
 
 export = () => {
-  let tsProject = makeTsProject({}, Config.TMP_DIR);
-  let toIgnore = readdirSync(Config.TMP_DIR)
+  const tsProject = makeTsProject({}, Config.TMP_DIR);
+  const toIgnore = readdirSync(Config.TMP_DIR)
     .filter(
       (f: string) =>
         lstatSync(join(Config.TMP_DIR, f)).isDirectory() &&
@@ -25,13 +25,13 @@ export = () => {
         '!' + join(Config.TMP_DIR, f, Config.NG_FACTORY_FILE + '.ts')
     );
 
-  let src = [
+  const src = [
     Config.TOOLS_DIR + '/manual_typings/**/*.d.ts',
     join(Config.TMP_DIR, '**/*.ts'),
     join(Config.TMP_DIR, `${Config.BOOTSTRAP_FACTORY_PROD_MODULE}.ts`),
     ...toIgnore
   ];
-  let result = gulp
+  const result = gulp
     .src(src)
     .pipe(plugins.plumber())
     .pipe(tsProject())
