@@ -20,10 +20,10 @@ export =
   class BuildJsTest extends TypeScriptTask {
     run() {
       let tsProject: any;
-      let typings = gulp.src( [
+      const typings = gulp.src( [
         Config.TOOLS_DIR + '/manual_typings/**/*.d.ts'
       ] );
-      let src = [
+      const src = [
         join(Config.APP_SRC, '**/*.spec.ts')
       ];
 
