@@ -1,18 +1,18 @@
 import * as util from 'gulp-util';
 import { join } from 'path';
 
-let fs = require('fs');
+const fs = require('fs');
 
 
 // if gulpfile.ts has been compiled then we need to rebuild the toolchain
 
 export = (done: any) => {
 
-  let checkFile = join(process.cwd(), 'tools', 'config.js');
+  const checkFile = join(process.cwd(), 'tools', 'config.js');
 
   // need to require the build.toolchain task as it won't be able to run after we run clear.files
-  let buildTools = require('./build.tools');
-  let cleanTools = require('./clean.tools');
+  const buildTools = require('./build.tools');
+  const cleanTools = require('./clean.tools');
 
   let rebuild = false;
 
@@ -32,7 +32,7 @@ export = (done: any) => {
     cleanTools();
 
     util.log('Running \'build.tools\' from check.tools');
-    let build = buildTools();
+    const build = buildTools();
 
     build.on('end', done);
 
