@@ -7,15 +7,15 @@ import Config from '../../config';
 
 class Protractor {
   server(port: number, dir: string) {
-    let app = express();
-    let root = resolve(process.cwd(), dir);
-    for (let proxy of Config.PROXY_MIDDLEWARE) {
+    const app = express();
+    const root = resolve(process.cwd(), dir);
+    for (const proxy of Config.PROXY_MIDDLEWARE) {
       app.use(proxy);
     }
     app.use(express.static(root));
     app.use(history('index.html', { root }));
     return new Promise((resolve, reject) => {
-      let server = app.listen(port, () => {
+      const server = app.listen(port, () => {
         resolve(server);
       });
     });
