@@ -4,8 +4,8 @@ import * as browserSync from 'browser-sync';
 import Config from '../../config';
 
 export = () => {
-  let coverageFolder = Config.COVERAGE_TS_DIR;
-  let watchedFiles: string[] = [join(coverageFolder, '**')];
+  const coverageFolder = Config.COVERAGE_TS_DIR;
+  const watchedFiles: string[] = [join(coverageFolder, '**')];
 
   // Serve files from the coverage of this project
   browserSync.create().init({
