@@ -10,7 +10,7 @@ const plugins = <any>gulpLoadPlugins();
  * Executes the build process, linting the TypeScript files using `codelyzer`.
  */
 export = () => {
-  let src = [
+  const src = [
     join(Config.APP_SRC, '**/*.ts'),
     '!' + join(Config.APP_SRC, '**/*.d.ts'),
       join(Config.E2E_SRC, '**/*.ts'),
